refactor: extract posting transport service
